requirements.txt
pandas==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dbt-postgres==1.7.0